Add UniqueHash trait
